tag: implicitly supply --list given the -n option
Change the "tag" command to treat the "-n" invocation as a list-like
option in addition to --contains, --points-at etc.

Most of the work for this was done in my earlier "tag: Implicitly
supply --list given another list-like option" commit, but I've split
off this patch since it's more contentious. Now these will be
synonymous:

    git tag -n 100
    git tag -n --list 100

Whereas before the former would die. This doesn't technically
introduce any more ambiguity than the aforementioned change applied to
th other list-like options, but it does introduce the possibility for
more confusion, since instead of the latter of these dying:

    git tag -n100
    git tag -n 100

It now works entirely differently, i.e. invokes list mode with a
filter for "100" as a pattern. I.e. it's synonymous with:

    git tag -n --list 100

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>